requirements.txt

Flask==2.0.2
torch==1.10.0
ultralytics==8.0.0
Pillow==8.4.0
requests==2.26.0
flask-cors==3.0.10
